backend/requirements.txt
django
django-rest-framework
whitenoise
pylint
djangorestframework_simplejwt